+++ gsa/src/web/pages/alerts/listpage.js
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-import React, {useEffect, useCallback} from 'react';
+import React, {useEffect, useCallback, useState} from 'react';
 
 import _ from 'gmp/locale';
 
@@ -24,10 +24,9 @@ import {ALERTS_FILTER_FILTER} from 'gmp/models/filter';
 import {hasValue} from 'gmp/utils/identity.js';
 
 import PropTypes from 'web/utils/proptypes';
-import withCapabilities from 'web/utils/withCapabilities.js';
+import useCapabilities from 'web/utils/useCapabilities.js';
 
 import EntitiesPage from 'web/entities/page.js';
-import withEntitiesContainer from 'web/entities/withEntitiesContainer.js';
 
 import ManualIcon from 'web/components/icon/manualicon.js';
 import NewIcon from 'web/components/icon/newicon.js';
@@ -42,31 +41,42 @@ import {createFilterDialog} from 'web/components/powerfilter/dialog.js';
 import DialogNotification from 'web/components/notification/dialognotification';
 import useDialogNotification from 'web/components/notification/useDialogNotification';
 
+import useReload from 'web/components/loading/useReload.js';
+import useDownload from 'web/components/form/useDownload.js';
+import Download from 'web/components/form/download.js';
+
 import {
   useLazyGetAlerts,
   useCloneAlert,
   useDeleteAlert,
   useTestAlert,
+  useDeleteAlertsByFilter,
+  useDeleteAlertsByIds,
+  useExportAlertsByFilter,
+  useExportAlertsByIds,
 } from 'web/graphql/alerts';
 
 import {
-  loadEntities,
-  selector as entitiesSelector,
-} from 'web/store/entities/alerts';
+  BulkTagComponent,
+  useBulkDeleteEntities,
+  useBulkExportEntities,
+} from 'web/entities/bulkactions.js';
+import {usePagination} from 'web/entities/usePagination.js';
+
 import usePageFilter from 'web/utils/usePageFilter.js';
 import useUserSessionTimeout from 'web/utils/useUserSessionTimeout.js';
 import useGmpSettings from 'web/utils/useGmpSettings.js';
 import usePrevious from 'web/utils/usePrevious.js';
 import useChangeFilter from 'web/utils/useChangeFilter.js';
 import useFilterSortBy from 'web/utils/useFilterSortby.js';
+import useSelection from 'web/utils/useSelection.js';
 
 import AlertComponent from './component.js';
 import AlertTable, {SORT_FIELDS} from './table.js';
-import useReload from 'web/components/loading/useReload.js';
-import {usePagination} from 'web/entities/usePagination.js';
 
-export const ToolBarIcons = withCapabilities(
-  ({capabilities, onAlertCreateClick}) => (
+export const ToolBarIcons = ({onAlertCreateClick}) => {
+  const capabilities = useCapabilities();
+  return (
     <IconDivider>
       <ManualIcon
         page="scanning"
@@ -77,8 +87,8 @@ export const ToolBarIcons = withCapabilities(
         <NewIcon title={_('New Alert')} onClick={onAlertCreateClick} />
       )}
     </IconDivider>
-  ),
-);
+  );
+};
 
 ToolBarIcons.propTypes = {
   onAlertCreateClick: PropTypes.func.isRequired,
@@ -91,6 +101,7 @@ const AlertFilterDialog = createFilterDialog({
 const AlertsPage = ({onChanged, onDownloaded, onError, ...props}) => {
   // Page methods and hooks
   const gmpSettings = useGmpSettings();
+  const [downloadRef, handleDownload] = useDownload();
   const [, renewSession] = useUserSessionTimeout();
   const [filter, isLoadingFilter] = usePageFilter('alert');
   const prevFilter = usePrevious(filter);
@@ -106,11 +117,18 @@ const AlertsPage = ({onChanged, onDownloaded, onError, ...props}) => {
     showError,
     showSuccess,
   } = useDialogNotification();
-
+  const {
+    selectionType,
+    selected = [],
+    changeSelectionType,
+    select,
+    deselect,
+  } = useSelection();
   const [sortBy, sortDir, handleSortChange] = useFilterSortBy(
     filter,
     changeFilter,
   );
+  const [tagsDialogVisible, setTagsDialogVisible] = useState(false);
 
   // Alert list state variables and methods
   const [
@@ -121,6 +139,15 @@ const AlertsPage = ({onChanged, onDownloaded, onError, ...props}) => {
   const [cloneAlert] = useCloneAlert();
   const [deleteAlert] = useDeleteAlert();
   const [testAlert] = useTestAlert();
+
+  const exportAlertsByFilter = useExportAlertsByFilter();
+  const exportAlertsByIds = useExportAlertsByIds();
+  const bulkExportAlerts = useBulkExportEntities();
+
+  const [deleteAlertsByIds] = useDeleteAlertsByIds();
+  const [deleteAlertsByFilter] = useDeleteAlertsByFilter();
+
+  const bulkDeleteAlerts = useBulkDeleteEntities();
 
   const timeoutFunc = useCallback(
     ({isVisible}) => {
@@ -161,6 +188,44 @@ const AlertsPage = ({onChanged, onDownloaded, onError, ...props}) => {
     alert => testAlert(alert.id).then(showSuccess, showError),
     [testAlert, showError, showSuccess],
   );
+
+  // Bulk action methods
+  const openTagsDialog = () => {
+    renewSession();
+    setTagsDialogVisible(true);
+  };
+
+  const closeTagsDialog = () => {
+    renewSession();
+    setTagsDialogVisible(false);
+  };
+
+  const handleBulkDeleteAlerts = () => {
+    return bulkDeleteAlerts({
+      selectionType,
+      filter,
+      selected,
+      entities: alerts,
+      deleteByIdsFunc: deleteAlertsByIds,
+      deleteByFilterFunc: deleteAlertsByFilter,
+      onDeleted: refetch,
+      onError: showError,
+    });
+  };
+
+  const handleBulkExportAlerts = () => {
+    return bulkExportAlerts({
+      entities: alerts,
+      selected,
+      filter,
+      resourceType: 'alerts',
+      selectionType,
+      exportByFilterFunc: exportAlertsByFilter,
+      exportByIdsFunc: exportAlertsByIds,
+      onDownload: handleDownload,
+      onError: showError,
+    });
+  };
 
   // Side effects
   useEffect(() => {
@@ -213,10 +278,13 @@ const AlertsPage = ({onChanged, onDownloaded, onError, ...props}) => {
             entities={alerts}
             entitiesCounts={counts}
             entitiesError={error}
+            entitiesSelected={selected}
+            filter={filter}
             filterEditDialog={AlertFilterDialog}
             filtersFilter={ALERTS_FILTER_FILTER}
             isLoading={isLoading}
             isUpdating={isLoading}
+            selectionType={selectionType}
             sectionIcon={<AlertIcon size="large" />}
             sortBy={sortBy}
             sortDir={sortDir}
@@ -230,6 +298,10 @@ const AlertsPage = ({onChanged, onDownloaded, onError, ...props}) => {
             onAlertEditClick={edit}
             onAlertTestClick={handleTestAlert}
             onAlertSaveClick={save}
+            onDeleteBulk={handleBulkDeleteAlerts}
+            onDownloadBulk={handleBulkExportAlerts}
+            onEntitySelected={select}
+            onEntityDeselected={deselect}
             onError={onError}
             onFilterChanged={changeFilter}
             onFilterCreated={changeFilter}
@@ -244,11 +316,24 @@ const AlertsPage = ({onChanged, onDownloaded, onError, ...props}) => {
             onLastClick={getLast}
             onNextClick={getNext}
             onPreviousClick={getPrevious}
+            onSelectionTypeChange={changeSelectionType}
+            onTagsBulk={openTagsDialog}
           />
           <DialogNotification
             {...notificationDialogState}
             onCloseClick={closeNotificationDialog}
           />
+          <Download ref={downloadRef} />
+          {tagsDialogVisible && (
+            <BulkTagComponent
+              entities={alerts}
+              selected={selected}
+              filter={filter}
+              selectionType={selectionType}
+              entitiesCounts={counts}
+              onClose={closeTagsDialog}
+            />
+          )}
         </React.Fragment>
       )}
     </AlertComponent>
@@ -261,9 +346,6 @@ AlertsPage.propTypes = {
   onError: PropTypes.func.isRequired,
 };
 
-export default withEntitiesContainer('alert', {
-  entitiesSelector,
-  loadEntities,
-})(AlertsPage);
+export default AlertsPage;
 
 // vim: set ts=2 sw=2 tw=80: